requirements.txt
# Panel Hopper - Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# Core BLE Communication
# =============================================================================
bleak>=0.21.0              # Bluetooth Low Energy library

# =============================================================================
# Image Processing
# =============================================================================
Pillow>=10.0.0             # Image manipulation
pilmoji>=2.0.0             # Emoji rendering for Pillow

# =============================================================================
# Web Interface
# =============================================================================
fastapi>=0.100.0           # Web framework
uvicorn>=0.23.0            # ASGI server
python-multipart>=0.0.6    # File uploads

# =============================================================================
# Configuration (used by vendor/bk_light)
# =============================================================================
PyYAML>=6.0                # YAML config support
